Fix unit tests + refactoring
