--- src/lib/binaryEncoder.ts
@@ -1,12 +1,0 @@
-/**
- * Convert a string to an Utf-8 ArrayBuffer
- */
-export function convertFromString(str: string): ArrayBuffer {
-    var buf = new ArrayBuffer(str.length);
-    var bufView = new Uint8Array(buf);
-    for (var i=0, strLen=str.length; i<strLen; i++) {
-        bufView[i] = str.charCodeAt(i);
-    }
-    return buf;
-}
-
